Implemented a semi-real shop
You can add items by editing UIShopScript.AvailableShopItems under the Shop1 object in the editor. Walk towards a shop to open it and press the shoot button (p for right now) to close it. Press the interact button (o) to "buy" the item. We do not have an inventory system yet so this hasn't been implemented yet.